refactor(DI-357): extract save logic from InfiniteDiscoveryArtworkCard into useInfiniteDiscoveryCardSave
Removes all onboarding-awareness from the card component. The card now
receives isSaved, handleSaveButtonPress, and handleDoubleTapSave from
the hook, with no knowledge of isNewUserOnboardingSession or the
onboarding artwork tracking logic.
